Finished Iterative Bruteforce Backtracking Algorithm
